fix: service and store layers
