+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -1,18 +1,27 @@
 package ru.yandex.practicum.filmorate.service;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exception.ObjectNotFoundException;
 import ru.yandex.practicum.filmorate.model.Film;
-import ru.yandex.practicum.filmorate.storage.FilmStorage;
+import ru.yandex.practicum.filmorate.storage.film.FilmStorage;
+import ru.yandex.practicum.filmorate.storage.film.RatingStorage;
 import java.util.Collection;
+import java.util.Map;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
+@Slf4j
 public class FilmService {
     private final FilmStorage filmStorage;
+    private final RatingStorage ratingStorage;
 
     @Autowired
-    public FilmService(FilmStorage filmStorage) {
+    public FilmService(FilmStorage filmStorage, RatingStorage ratingStorage) {
         this.filmStorage = filmStorage;
+        this.ratingStorage = ratingStorage;
     }
 
     public Collection<Film> findAll() {
@@ -20,26 +29,50 @@ public class FilmService {
     }
 
     public Film findById(Integer id) {
-        return filmStorage.findById(id);
+        Optional<Film> result = filmStorage.findById(id);
+
+        if (result.isEmpty()) {
+            throw new ObjectNotFoundException("Film with id: " + id + " not found!");
+        }
+
+        return result.get();
     }
 
     public Collection<Film> findPopular(int count) {
-        return filmStorage.findPopular(count);
+        return ratingStorage.findAll().entrySet()
+                .stream()
+                .sorted((x1, x2) -> x2.getValue().size() - x1.getValue().size())
+                .limit(count)
+                .map(Map.Entry::getKey)
+                .map(filmStorage::findById)
+                .filter(Optional::isPresent)
+                .map(Optional::get)
+                .collect(Collectors.toList());
     }
 
     public Film create(Film film) {
-        return filmStorage.create(film);
+        Film result = filmStorage.create(film);
+        ratingStorage.createRatingContainer(result.getId());
+
+        return result;
     }
 
     public Film update(Film film) {
-        return filmStorage.update(film);
+        Optional<Film> result = filmStorage.update(film);
+        if (result.isEmpty()) {
+            throw new ObjectNotFoundException("Film with id: " + film.getId() + " not found!");
+        }
+
+        return result.get();
     }
 
     public void addLike(Integer filmId, Integer userId) {
-        filmStorage.addLike(filmId, userId);
+        ratingStorage.addLike(filmId, userId);
+        log.debug("Film with id:{} was liked by user with id:{}", filmId, userId);
     }
 
     public void removeLike(Integer filmId, Integer userId) {
-        filmStorage.removeLike(filmId, userId);
+        ratingStorage.removeLike(filmId, userId);
+        log.debug("Film with id:{} was unliked by user with id:{}", filmId, userId);
     }
 }
