+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -1,20 +1,25 @@
 package ru.yandex.practicum.filmorate.service;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.model.User;
-import ru.yandex.practicum.filmorate.storage.UserStorage;
+import ru.yandex.practicum.filmorate.storage.user.FriendStorage;
+import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 import java.util.Collection;
 import java.util.HashSet;
 import java.util.Set;
 
 @Service
+@Slf4j
 public class UserService {
     private final UserStorage userStorage;
+    private final FriendStorage friendStorage;
 
     @Autowired
-    public UserService(UserStorage userStorage) {
+    public UserService(UserStorage userStorage, FriendStorage friendStorage) {
         this.userStorage = userStorage;
+        this.friendStorage = friendStorage;
     }
 
     public Collection<User> findAll() {
@@ -26,7 +31,7 @@ public class UserService {
     }
 
     public Collection<User> findFriends(int userId) {
-        return userStorage.findFriends(userId);
+        return friendStorage.findFriends(userId);
     }
 
     public User createUser(User user) {
@@ -38,22 +43,36 @@ public class UserService {
     }
 
     public void createFriendship(Integer userId1, Integer userId2) {
-        userStorage.addFriend(userId1, userId2);
-        userStorage.addFriend(userId2, userId1);
+        addFriend(userId1, userId2);
+        addFriend(userId2, userId1);
     }
 
     public void removeFriendship(Integer userId1, Integer userId2) {
-        userStorage.removeFriend(userId1, userId2);
-        userStorage.removeFriend(userId2, userId1);
+        removeFriend(userId1, userId2);
+        removeFriend(userId2, userId1);
     }
 
     public Collection<User> findCommonFriends(Integer firstUserId, Integer secondUserId) {
-        Collection<User> firstUserFriends = userStorage.findFriends(firstUserId);
-        Collection<User> secondUserFriends = userStorage.findFriends(secondUserId);
+        Collection<User> firstUserFriends = friendStorage.findFriends(firstUserId);
+        Collection<User> secondUserFriends = friendStorage.findFriends(secondUserId);
 
         Set<User> commonFriends = new HashSet<>(firstUserFriends);
         commonFriends.retainAll(secondUserFriends);
 
         return commonFriends;
+    }
+
+    private void addFriend(int userId, int friendId) {
+        User friend = userStorage.findById(friendId);
+        friendStorage.addFriend(userId, friend);
+
+        log.debug("User with id:{} got a friend with id:{}", userId, friendId);
+    }
+
+    private void removeFriend(int userId, int friendId) {
+        User friend = userStorage.findById(friendId);
+        friendStorage.removeFriend(userId, friend);
+
+        log.debug("User with id:{} lost a friend with id:{}", userId, friendId);
     }
 }
